create FileClient
