+++ bin/reth/src/chain/init.rs
@@ -1,0 +1,10 @@
+use clap::Parser;
+
+#[derive(Debug, Parser)]
+pub struct InitCommand;
+
+impl InitCommand {
+    pub async fn execute(&self) -> eyre::Result<()> {
+        todo!()
+    }
+}
